gmail_api/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
boto3
sqlalchemy
pymysql
python-dotenv 
pyodbc
pymssql